requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.6
django-toolbelt==0.0.1
djangorestframework==3.6.2
gunicorn==19.7.0
psycopg2==2.7
requests==2.13.0
static3==0.7.0
